Inpainter-Training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
huggingface-hub>=0.16.0
peft>=0.4.0
pyarrow>=12.0.0
wandb>=0.15.0
Pillow>=9.0.0
numpy>=1.24.0
